Fix writing of PDFs larger than 2GB
Represent file offsets as `long` instead of `int` to prevent overflows.

There's also another issue after 9,999,999,999 bytes: the old-style xref
table only allots 10 digits for each object's offset. In this case, we
must write an xref stream instead.

+++ openpdf/src/main/java/com/lowagie/text/pdf/OutputStreamCounter.java
@@ -55,7 +55,7 @@ import java.io.OutputStream;
 public class OutputStreamCounter extends OutputStream {
     
     protected OutputStream out;
-    protected int counter = 0;
+    protected long counter = 0;
     
     /** Creates a new instance of OutputStreamCounter */
     public OutputStreamCounter(OutputStream out) {
@@ -160,7 +160,7 @@ public class OutputStreamCounter extends OutputStream {
         out.write(b, off, len);
     }
     
-    public int getCounter() {
+    public long getCounter() {
         return counter;
     }
